Adds Script to Auto Convert Input/Output Files to UNIX
If users download these files onto a windows computer,
there is a good chance the file type will be changed to
CRLF. This way, we ensure that they do not experience
said issue.

Note: I have not tested this, as I do not use the scripts.
If someone could make sure this works as intented,
that would be great.